Feature/category services

+++ eCommerce.BackendApi/Services/CategoryService.cs
@@ -1,6 +1,4 @@
-﻿using System;
-using System.Net.Http.Headers;
-using eCommerce.BackendApi.Data.EF;
+﻿using eCommerce.BackendApi.Data.EF;
 using eCommerce.BackendApi.Interfaces;
 using eCommerce.BackendApi.Models;
 using eCommerce.Shared.ViewModels.Categories;
@@ -12,7 +10,6 @@ namespace eCommerce.BackendApi.Services
     {
         private readonly ApplicationDbContext _dbContext;
         private readonly IFileStorageService _fileStorageService;
-        private const string FILE_SOURCE_FOLDER_NAME = "file-source";
 
         public CategoryService(ApplicationDbContext dbContext, IFileStorageService fileStorageService)
         {
@@ -72,37 +69,55 @@ namespace eCommerce.BackendApi.Services
 
             if (req.Image != null)
             {
-                category.ImageUrl = await SaveFile(req.Image);
+                category.ImageUrl = await _fileStorageService.SaveFile(req.Image);
             }
             _dbContext.Categories.Add(category);
             await _dbContext.SaveChangesAsync();
             return category.Id;
         }
 
-        public async Task<int> DeleteCategory(int categoryId)
+        public async Task<int> DeleteCategory(CategoryDeleteRequest req)
         {
-            var category = await _dbContext.Categories.FindAsync(categoryId);
+            var category = await _dbContext.Categories.FindAsync(req.Id);
 
             if (category == null)
             {
-                throw new Exception($"Cannot delete category because CategoryID not found");
+                throw new Exception($"Cannot delete category because CategoryID {req.Id} is null not found");
             }
 
-            await _fileStorageService.DeleteFileAsync(category.ImageUrl);
+            if(category.ImageUrl != null)
+            {
+                await _fileStorageService.DeleteFileAsync(category.ImageUrl);
+            }
 
             _dbContext.Categories.Remove(category);
             return await _dbContext.SaveChangesAsync();
         }
 
-        private async Task<string> SaveFile(IFormFile file)
+        public async Task<int> UpdateCategory(CategoryUpdateRequest req)
         {
-            #pragma warning disable CS8602 // Dereference of a possibly null reference.
-            var originalFileName = ContentDispositionHeaderValue.Parse(file.ContentDisposition)
-                                    .FileName.Trim('"');
-            #pragma warning restore CS8602 // Dereference of a possibly null reference.
-            var fileName = $"{Guid.NewGuid()}{Path.GetExtension(originalFileName)}";
-            await _fileStorageService.SaveFileAsync(file.OpenReadStream(), fileName);
-            return FILE_SOURCE_FOLDER_NAME + "/" + fileName;
+            var category = await _dbContext.Categories.FindAsync(req.Id);
+
+            if(category == null)
+            {
+                throw new Exception($"Cannot update category because CategoryID {req.Id} is null or not found");
+            }
+
+            if(req.Name != null)
+            {
+                category.Name = req.Name;
+            }
+
+            category.Description = req.Description;
+            category.ParentId = req.ParentId;
+
+            if(req.Image != null)
+            {
+                category.ImageUrl = await _fileStorageService.SaveFile(req.Image);
+            }
+
+            _dbContext.Categories.Update(category);
+            return await _dbContext.SaveChangesAsync();
         }
